+++ assets/ui/src/lib/api/client.ts
@@ -592,6 +592,7 @@ export const api = {
       value_list?: string[];
       value_name_0?: string;
       value_name_1?: string;
+      channel_address?: string;
     },
     central: string,
   ) {
@@ -615,6 +616,7 @@ export const api = {
       value_name_1?: string;
       is_visible?: boolean;
       is_logged?: boolean;
+      channel_address?: string;
     },
     central?: string,
   ) {